+++ peacecorps/paygov/models.py
@@ -1,18 +1,1 @@
-from datetime import timedelta
-
-from django.conf import settings
-from django.db import models
-from django.utils import timezone
-
-
-def default_expire_time():
-    return timezone.now() + timedelta(minutes=settings.DONOR_EXPIRE_AFTER)
-
-
-class DonorInfo(models.Model):
-    """Represents a blob of donor information which will be requested by
-    pay.gov. We need to limit accessibility as it contains PII"""
-    agency_tracking_id = models.CharField(max_length=21, primary_key=True)
-    fund = models.ForeignKey('peacecorps.Fund', related_name='donorinfos')
-    xml = models.TextField()    # @todo: encrypt
-    expires_at = models.DateTimeField(default=default_expire_time)
+# This file should contain no models that are needed by other apps
