+++ peacecorps/paygov/views.py
@@ -5,7 +5,7 @@ from django.http import HttpResponseNotAllowed
 from django.shortcuts import get_object_or_404
 from django.views.decorators.csrf import csrf_exempt
 
-from paygov.models import DonorInfo
+from peacecorps.models import DonorInfo
 
 
 @csrf_exempt
